[UI] Work on splitting the extraction worker up more.
This commit also turned off PCHing for the UI, as it actually *hurts* compile time a bit. Oops!

+++ src/core/tegra_texture.cpp
@@ -170,7 +170,7 @@ namespace core {
 			GetBlockWidth(),
 			GetRealBlockHeight(GetBlockHeight()),
 
-			std::log2(GetRealBlockHeight(GetBlockHeight()))
+			(uint32)std::log2(GetRealBlockHeight(GetBlockHeight()))
 		};
 
 		// Create a copy of the texture data that we will work with.
